# 5-exam/gateway/api/swagger/docs/swagger.yaml
definitions:
  booking.BookingInfo:
    properties:
      checkInDate:
        type: string
      checkOutDate:
        type: string
      hotel_id:
        type: string
      roomType:
        type: string
      totalAmount:
        type: number
      totalDays:
        type: integer
      user_id:
        type: string
    type: object
  booking.BookingResponse:
    properties:
      booking_id:
        type: string
      checkInDate:
        type: string
      checkOutDate:
        type: string
      hotel_id:
        type: string
      roomType:
        type: string
      status:
        type: string
      totalAmount:
        type: number
      totalDays:
        type: integer
      user_id:
        type: string
    type: object
  booking.Response:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This swagger UI was created to manage hotel bookings
  title: 'Project: Hotel Booking API'
  version: "1.0"
paths:
  /api/bookings:
    post:
      consumes:
      - application/json
      description: This method used  to create new  booking
      parameters:
      - description: Role
        in: header
        name: role
        required: true
        type: string
      - description: Booking Details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/booking.BookingInfo'
      produces:
      - application/json
      responses:
        "201":
          description: Booking  Confirmation
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "400":
          description: Unable to read booking  details !
          schema: {}
        "403":
          description: Unauthorized access
          schema: {}
        "500":
          description: Unable to get response
          schema: {}
      security:
      - BearerAuth: []
      summary: Creates New Booking
      tags:
      - BOOKINGS
  /api/bookings/{id}:
    delete:
      consumes:
      - application/json
      description: This method used  to delete   booking detail
      parameters:
      - description: Role
        in: header
        name: role
        required: true
        type: string
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Deleted Succesfully
          schema:
            $ref: '#/definitions/booking.Response'
        "403":
          description: Unauthorized access
          schema: {}
        "500":
          description: Unable to get response
          schema: {}
      security:
      - BearerAuth: []
      summary: Deletes  Booking Details
      tags:
      - BOOKINGS
    get:
      consumes:
      - application/json
      description: This method used  to get   booking detail
      parameters:
      - description: Role
        in: header
        name: role
        required: true
        type: string
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Deleted Succesfully
          schema:
            $ref: '#/definitions/booking.Response'
        "403":
          description: Unauthorized access
          schema: {}
        "500":
          description: Unable to get response
          schema: {}
      security:
      - BearerAuth: []
      summary: gets  Booking Details By ID
      tags:
      - BOOKINGS
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"